Reorganize code in multiple files
