+++ perf_tool/src/utils.hpp
@@ -1,0 +1,28 @@
+// Copyright 2022 Open Source Robotics Foundation, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef PERF_TOOL__UTILS_HPP_
+#define PERF_TOOL__UTILS_HPP_
+
+#include <string>
+
+#include "rmw/types.h"
+
+namespace perf_tool
+{
+std::string
+stringify_gid(const rmw_gid_t & rmw_gid);
+}
+
+#endif  // PERF_TOOL__UTILS_HPP_
